added protocols. improved summary keywords. fixed file bugs

+++ bids_constructor.py
@@ -4,7 +4,8 @@ import os
 import re
 from pathlib import Path
 import pandas as pd
-import nibabel as nib
+
+from criteria import inclusion_or_exclusion_criteria
 
 def check_path(path):
     if not os.path.isdir(path):
@@ -63,48 +64,6 @@ def get_nifti_info(nifti_file, json_data):
         info_nifti.num_id = '0'+info_nifti.num_id
     return path, name, info_nifti
 
-def inclusion_or_exclusion_criteria(files, info_files, bids_code):
-    """
-    Define here any criteria you want to consider in order for the BIDS conversion to proceed or not.
-
-    Inputs
-    --------
-    files (list): Contains the files outputted by the dcm2niix software
-    info_files (object): Contains information on the batch of files converted - Obtained from the function get_nifti_info
-    bids_code (dict): Contains information about the coding keys to transform to bids
-
-    Outputs
-    --------
-    proceed (bool): Encoding whether criteria is met for continuation of the conversion
-    mri (str, None): Encodes the type of mri protocol found. 
-                     If None means that no protocol is of interest
-    """
-    
-    # CONSIDER ADDING (from) HERE ANY OTHER PROTOCOLS YOU MIGHT NEED OR WANT TO SEARCH 
-    mri = re.match(r"dti|MPR|MPRAGE|BOLD|t1_mprage|t2_spc_1mm_p2", info_files.protocol) 
-    if not mri == None:
-        mri = mri.group()
-    
-        # CONSIDER ADDING ANY OTHER CRITERIA PRESENT IN YOUR DATASET
-        if bids_code[mri] == 'dwi':
-            series = json.load(open([f for f in files if '.json' in f][0], 'r'))["SeriesDescription"]
-            if ('FA' in series) or ('ADC' in series) or ('TENSOR' in series):
-                proceed = False
-            else:
-                proceed = True
-        elif bids_code[mri] == 'bold':
-            data = nib.load([f for f in files if '.nii' in f][0]).get_fdata()
-            if data.shape[-1] >= 100:
-                proceed = True
-            else:
-                proceed = False
-        else:
-            proceed = True
-    else:
-        proceed = False
-    
-    return proceed, mri
-
 def organize_niftis(niftis, root_subject, root_name, mri):
     # Moving niftis
     for nifti in niftis:
@@ -116,16 +75,23 @@ def organize_niftis(niftis, root_subject, root_name, mri):
         backup = 0
         if os.path.exists(full_name):
             check_path(root_subject+mri+'/backup/')
-            full_name = root_subject+mri+'/backup/'+root_name+'_bck.'+extension
-            backup = 1
+            full_name = root_subject+mri+'/backup/'+root_name+'_bck-'+str(backup)+'.'+extension
+            backup += 1
         Path(nifti).rename(full_name)
+
     return backup
 
 def convert_dicom_session(f, config, bids_code):
     dicom_folders, num_dicoms = get_folders(path=f, search_type='directory')
-
+    backup_anat, backup_dwi, backup_func = 0, 0, 0
+    
     ##### Convert and process each DICOM session ####
-    current_session = {"session_id": '', "acq_time": '',"FOLDER": f.split("/")[-1],  "anat": 0, "dwi": 0, "func": 0, "BACKUP": 0}
+    current_session = {
+        "session_id": '', "acq_time": '',"FOLDER": f.split("/")[-1],
+        "anat": 0, "dwi": 0, "func": 0,
+        "BACKUP-anat": 0, "BACKUP-dwi": 0, "BACKUP-func": 0
+    }
+
     for df in dicom_folders:
         output = Popen(
             f"dcm2niix -o {config['data']['output_path']} -f %n--%p--%t -z {'y' if config['data']['gzip'] else 'n'} {df}", shell=True, stdout=PIPE
@@ -168,31 +134,50 @@ def convert_dicom_session(f, config, bids_code):
                 if new:
                     bids_tree(subject_folder, mris=config['subjects']['mris'])
 
-                ## Organize nifti files ##   
-                backup = organize_niftis(niftis, subject_folder, file_name, mri)
+                if mri in config['subjects']['mris']:
+                    ## Organize nifti files ##   
+                    backup = organize_niftis(niftis, subject_folder, file_name, mri)
+                    if mri == 'anat':
+                        backup_anat = backup
+                    elif mri == 'dwi':
+                        backup_dwi = backup
+                    else:
+                        backup_func = backup
 
-                ## Update intra-session MRI ##
-                current_session[mri] += 1
+                    ## Update intra-session MRI ##
+                    current_session[mri] += 1
+                else:
+                    # Remove converted files
+                    for nf in niftis:
+                        os.remove(nf)  
+
             else:
                 # Remove converted files
                 for nf in niftis:
                     os.remove(nf)    
     
     #### Update Subject Summary ####
-    #   Current Session
-    current_session['acq_time'] = info_niftis.time
-    current_session['session_id'] = bids_code[info_niftis.session]["session"]
-    current_session['BACKUP'] = backup
-    tsv_name = config['data']['output_path'] + 'sub-' + bids_code[info_niftis.session]["ID"] + info_niftis.num_id + '/' + \
-        'sub-' + bids_code[info_niftis.session]["ID"] + info_niftis.num_id + '_sessions.tsv'
-    try: #   Non-available Summary from previous folders
-        subject_summary = pd.read_csv(tsv_name, sep='\t')
-    except: # Available Summary from previous folders
-        subject_summary = pd.DataFrame(columns=['session_id', 'acq_time', 'FOLDER', 'anat', 'dwi', 'func', 'BACKUP'])
-    #   Combine earlier sessions with current
-    current_session = pd.DataFrame([current_session], columns=['session_id', 'acq_time', 'FOLDER', 'anat', 'dwi', 'func', 'BACKUP'])
-    subject_summary = pd.concat([subject_summary, current_session], ignore_index=True)
-    subject_summary.to_csv(tsv_name, sep='\t', index=False)
+    if os.path.exists(config['data']['output_path'] + 'sub-' + bids_code[info_niftis.session]["ID"] + info_niftis.num_id + '/'):
+        #   Current Session
+        current_session['acq_time'] = info_niftis.time
+        current_session['session_id'] = bids_code[info_niftis.session]["session"]
+        current_session['BACKUP-anat'] = backup_anat
+        current_session['BACKUP-dwi'] = backup_dwi
+        current_session['BACKUP-func'] = backup_func
+        tsv_name = config['data']['output_path'] + 'sub-' + bids_code[info_niftis.session]["ID"] + info_niftis.num_id + '/' + \
+            'sub-' + bids_code[info_niftis.session]["ID"] + info_niftis.num_id + '_sessions.tsv'
+        try: #   Non-available Summary from previous folders
+            subject_summary = pd.read_csv(tsv_name, sep='\t')
+        except: # Available Summary from previous folders
+            subject_summary = pd.DataFrame(
+                columns=['session_id', 'acq_time', 'FOLDER', 'anat', 'dwi', 'func', 'BACKUP-anat', 'BACKUP-dwi', 'BACKUP-func']
+            )
+        #   Combine earlier sessions with current
+        current_session = pd.DataFrame(
+            [current_session], columns=['session_id', 'acq_time', 'FOLDER', 'anat', 'dwi', 'func', 'BACKUP-anat', 'BACKUP-dwi', 'BACKUP-func']
+        )
+        subject_summary = pd.concat([subject_summary, current_session], ignore_index=True)
+        subject_summary.to_csv(tsv_name, sep='\t', index=False)
         
 if __name__ == '__main__':
     pass
